% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEOF.R
\name{addEOF}
\alias{addEOF}
\title{add EOF comment}
\usage{
addEOF(path = ".", filetypes = c("inc", "prn", "gms"))
}
\arguments{
\item{path}{path to the main folder of the model}

\item{filetypes}{file types the function should be applied to}
}
\description{
Add EOF text to all files in the code in order to ease debugging
}
\examples{
\dontrun{
addEOF()
}

}
\seealso{
\code{\link{removeEOF}}
}
\author{
Anastasis Giannousakis
}
