% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGitHubActions.R
\name{addGitHubActions}
\alias{addGitHubActions}
\title{addGitHubActions}
\usage{
addGitHubActions(lib = ".", config = NULL)
}
\arguments{
\item{lib}{Path to the package}

\item{config}{The build configuration loaded from .buildlibrary. If NULL, loads it automatically.}
}
\description{
This function adds a standard Github action workflow called "check.yaml" to the project which runs
lucode2::check(), checks the validation key, and creates a coverage report using codecov. This file is overwritten
automatically each time this function is run and should not be edited by hand.
}
\details{
#' If the config option `UsePkgDown` is set to TRUE in the `.buildlibrary` file,
this function also adds a pkgdown.yaml workflow that builds and deploys package documentation
to GitHub Pages, and creates a default `_pkgdown.yml` configuration file if it doesn't exist.

In addition, this function adds a codecov.yml to the repository, if not already existing. This file
is only created if missing and can be edited manually.
}
\examples{
\dontrun{
addGitHubActions()
}
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich
}
