% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFormat.R
\name{autoFormat}
\alias{autoFormat}
\title{autoFormat}
\usage{
autoFormat(
  files = getFilesToLint(),
  ignoreLintFreeFiles = TRUE,
  lintAfterwards = TRUE,
  formatTF = FALSE
)
}
\arguments{
\item{files}{A character vector of paths to files that should be auto-formatted.}

\item{ignoreLintFreeFiles}{If set to TRUE (the default) files without linter warnings are not auto-formatted.}

\item{lintAfterwards}{If set to TRUE (the default) return linter results for the auto-formatted files.}

\item{formatTF}{If set to TRUE (the default is FALSE) ask for T and F to be converted into TRUE and FALSE.
Note that it currently can't find out reliably if T or F are aliases for TRUE and FALSE, so manual intervention
is needed.}
}
\description{
Apply auto-formatting using styler::style_file to the given files. Does not change indentation.
}
\examples{
\dontrun{
lucode2::autoFormat()
}
}
\seealso{
\code{\link{getFilesToLint}}
}
\author{
Pascal Führlich
}
