% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_versions.R
\name{check_versions}
\alias{check_versions}
\title{Package version check tool}
\usage{
check_versions(mail = TRUE, test = FALSE, gitpath = NULL)
}
\arguments{
\item{mail}{whether an email notification is sent to RSE}

\item{test}{to test whether auto email sending works}

\item{gitpath}{if an email notification has to be sent out, the path to the git repo}
}
\description{
Checks if there are CRAN-packages with the same name as those in our PIK-CRAN whose version is newer
}
\author{
Anastasis Giannousakis
}
