% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDeps.R
\name{checkDeps}
\alias{checkDeps}
\title{checkDeps}
\usage{
checkDeps(
  descriptionFile = ".",
  dependencyTypes = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{descriptionFile}{Path to a DESCRIPTION or a path that belongs to a source package project.}

\item{dependencyTypes}{The types of depencies to check. Must be a
subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances").}
}
\value{
Invisibly, a named logical vector indicating whether each package requirement is met.
}
\description{
Check if package requirements specified in the given DESCRIPTION are met.
}
\details{
A warning is thrown for each required package that is not installed, and
each installed package whose version number is lower than what is required.
Only ">=" version requirements are supported.
}
\examples{
checkDeps(system.file("DESCRIPTION", package = "lucode2"))
}
\author{
Pascal Führlich
}
