% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRepoUpToDate.R
\name{checkRepoUpToDate}
\alias{checkRepoUpToDate}
\title{Check if repo is up to date with upstream}
\usage{
checkRepoUpToDate(pathToRepo = ".", autoCheckRepoUpToDate = TRUE)
}
\arguments{
\item{pathToRepo}{The path to the git repo.}

\item{autoCheckRepoUpToDate}{If FALSE do not check automatically and instead
just ask the user.}
}
\description{
Checks whether the local repo is up-to-date with the remote tracking branch
and the default branch of the upstream remote. Will throw an error if not
up-to-date and prints a git command to update.
}
\author{
Pascal Führlich
}
