% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRequiredPackages.R
\name{checkRequiredPackages}
\alias{checkRequiredPackages}
\title{checkRequiredPackages}
\usage{
checkRequiredPackages(
  requiredPackages,
  requiredFor = "",
  installFunction = install.packages,
  readlineFunction = readline
)
}
\arguments{
\item{requiredPackages}{One or more names of packages that are checked using requireNamespace.}

\item{requiredFor}{Optional single character. What the packages are required for, usually the name of a function.}

\item{installFunction}{Optional function, default NULL means install.packages. Will be called during
checkRequiredPackages like this: installFunction(missingPackages). Only needed if the required packages are not
available on CRAN or the configured repos (getOption("repos")).}

\item{readlineFunction}{Only used for testing. A function to get an answer from the user.}
}
\description{
Check if one or more packages are available and try to install the missing packages. Throw an error if at least one
of the required packages is still missing after the installation attempt.
}
\examples{
\dontrun{
checkRequiredPackages(c("ggplot2", "lusweave"), "lucode2::readRuntime(..., plot = TRUE)")
}
}
\seealso{
\code{\link{requireNamespace}}
}
\author{
Pascal Führlich
}
