% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkup.R
\name{checkup}
\alias{checkup}
\title{checkup}
\usage{
checkup()
}
\value{
Invisibly, the report as a list.
}
\description{
Checks the current R ecosystem for common problems and reports info such as OS and R version.
}
\author{
Pascal Führlich
}
