% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eprint_list.R
\name{eprint_list}
\alias{eprint_list}
\title{Extended list print}
\usage{
eprint_list(var_list, envir = parent.frame())
}
\arguments{
\item{var_list}{Vector containing names of several variables that should be
printed}

\item{envir}{environment from which the variable should be read (by default
the environment from which the function is called)}
}
\description{
Same as \code{\link{eprint}}, but expecting a vector with variable names
}
\examples{

a <- 1:3
b <- "blub"
lucode2:::eprint_list(c("a", "b"))
}
\seealso{
\code{\link{eprint}}
}
\author{
Jan Philipp Dietrich
}
