% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCoupledruns.R
\name{findCoupledruns}
\alias{findCoupledruns}
\title{findCoupledruns}
\usage{
findCoupledruns(resultsfolder)
}
\arguments{
\item{resultsfolder}{Path to an output folder.}
}
\value{
A vector containing the names of the scenarios.
}
\description{
Extracts scenario names from coupled runs in the given outputfolder. The
scenario names will be extracted based on the folder names of the results
folders.
}
\author{
David Klein
}
