% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDeps.R
\name{findDeps}
\alias{findDeps}
\title{findDeps}
\usage{
findDeps(devDeps = TRUE)
}
\arguments{
\item{devDeps}{Whether development dependencies should also be checked.}
}
\value{
A dataframe documenting which dependency is needed where.
}
\description{
Find all dependencies of an R project.
}
\details{
This is a wrapper around `renv::dependencies()` that does not report
dependencies on core R packages, because these are always available.
}
\seealso{
\code{\link[renv]{dependencies}}
}
\author{
Pascal Sauer
}
