% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIterations.R
\name{findIterations}
\alias{findIterations}
\title{findIterations}
\usage{
findIterations(runname, modelpath = ".", latest = FALSE)
}
\arguments{
\item{runname}{Scenarioname or vector of scenarionames.}

\item{modelpath}{Path or vector of paths where iterations are searched for.}

\item{latest}{Logical indicating if only the latest iteration of a runname
is returned.}
}
\value{
A vector containing the paths to the iterations of coupled runs.
}
\description{
Collects paths to all coupled runs (iterations) in \code{modelpath} that
contain \code{runname}. For each entry in \code{runname} the paths are
sorted by the modification time of the respective fulldata.gdx
}
\author{
David Klein
}
