% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixBuildLibraryMergeConflict.R
\name{fixBuildLibraryMergeConflict}
\alias{fixBuildLibraryMergeConflict}
\title{fixBuildLibraryMergeConflict}
\usage{
fixBuildLibraryMergeConflict(lib = ".")
}
\arguments{
\item{lib}{The path to the project with a merge conflict.}
}
\description{
Fix merge conflicts in files auto-edited by buildLibrary (.buildlibrary, .zenodo.json, DESCRIPTION, README.md).
}
\details{
The ValidationKey in .buildlibrary will be set to an empty string and the higher version number
in DESCRIPTION is used. buildLibrary needs to be run after this function to deal with the
ValidationKey and the merge markers in .zenodo and README.md.
}
