% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixfile.R
\name{fixfile}
\alias{fixfile}
\title{fixfile}
\usage{
fixfile(f)
}
\arguments{
\item{f}{path to the file that should be fixed}
}
\description{
Fix file format of text files (e.g. line/file endings).
}
\author{
Jan Philipp Dietrich
}
