% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTF.R
\name{formatTF}
\alias{formatTF}
\title{formatTF}
\usage{
formatTF(pathToFile)
}
\arguments{
\item{pathToFile}{Path to the R source file where T and F should be renamed.}
}
\description{
Convert T and F to TRUE and FALSE, respectively. The tool finds T and F and asks about
each ocurrence if it should be converted. Please check carefully before agreeing to the
conversion, there are false positives e.g. if a dataframe column name is T or F.
}
\examples{
\dontrun{
lucode2::formatTF("myFile.R")
}
}
\seealso{
\code{\link{autoFormat}}, \code{\link{snakeToCamel}}
}
\author{
Pascal Führlich
}
