% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionHeaderDefaults.R
\name{functionHeaderDefaults}
\alias{functionHeaderDefaults}
\title{functionHeaderDefaults}
\usage{
functionHeaderDefaults(...)
}
\arguments{
\item{...}{parameters that shall be assigned to the global environment of R}
}
\value{
no direct return, values are assigned to .GlobalEnv
}
\description{
used to quickly read in the default values of a function
}
\examples{
\dontrun{
test <- function(a = "klk", b = "kjlkv", kk = 3) {
paste(a, b, kk)
}
functionHeaderDefaults(a = "klk", b = "kjlkv", kk = 3)
print(a)
paste(a, b, kk)
}

}
\author{
Benjamin Leon Bodirsky
}
