% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterLoad.R
\name{getClusterLoad}
\alias{getClusterLoad}
\title{getClusterLoad}
\usage{
getClusterLoad()
}
\value{
NULL, if command "sclass" is not available,
otherwise returns a named vector with current load on
available partitions
}
\description{
Returns information about cluster load in case that
the command "sclass" is available. Otherwise, it
returns NULL
}
\author{
Jan Philipp Dietrich
}
