% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilesToLint.R
\name{getFilesToLint}
\alias{getFilesToLint}
\title{getFilesToLint}
\usage{
getFilesToLint(pathToGitRepo = ".")
}
\arguments{
\item{pathToGitRepo}{path to a git repository}
}
\description{
Get the R files the current git user is responsible for to pass them to the auto-formatter and/or linter.
}
\details{
The files of interest are identified using git via system() (and shell() for windows). All currently untracked files
and changed files (both staged and unstaged) are collected, as well as files that were changed in non-merge commits
authored by the current git user since the last version commit (a commit where .buildLibrary was changed, presumably
to increase the version number). Of those the absolute paths to .R, .Rmd and .Rnw files are returned as a character
vector.
}
\examples{
lucode2:::getFilesToLint()
}
\seealso{
\code{\link{lint}}, \code{\link{autoFormat}}
}
\author{
Pascal Sauer
}
