% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageAuthors.R
\name{getPackageAuthors}
\alias{getPackageAuthors}
\title{getPackageAuthors}
\usage{
getPackageAuthors(folder = "R")
}
\arguments{
\item{folder}{R folder of the package}
}
\description{
Creates a suggestion for an Authors entry for the DESCRIPTION
of a package. Suggestion is based on author information in roxygen
headers and authors are ranked based on number of mentionings.
}
\details{
Please be aware that the output will most likely require some
manual processing before it can be used in the DESCRIPTION!
}
\author{
Jan Philipp Dietrich
}
