% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScenNames.R
\name{getScenNames}
\alias{getScenNames}
\title{getScenNames}
\usage{
getScenNames(dirs)
}
\arguments{
\item{dirs}{vector of paths to the used output folders.}
}
\value{
A vector containing the titles used as scenario names for e.g. plots
}
\description{
Extract module information of a GAMS model.
}
\author{
Lavinia Baumstark
}
