% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitAuthors.R
\name{gitAuthors}
\alias{gitAuthors}
\title{gitAuthors}
\usage{
gitAuthors()
}
\value{
Invisibly, a data.frame with the following columns: raw, firstName, lastName, mailAdress.
}
\description{
Print each git author's first name, last name, and mail address in a way that could be copy pasted
into DESCRIPTION. Please note that this list usually still needs to be carefully checked
e.g. for duplicates and cleaned.
}
\author{
Pascal Sauer
}
