% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubMetadataExtract.R
\name{githubMetadataExtract}
\alias{githubMetadataExtract}
\title{githubMetadataExtract}
\usage{
githubMetadataExtract(org)
}
\arguments{
\item{org}{name of the organization from which the metadata should be extracted}
}
\description{
A function to extract metadata information from a GitHub organization, summarizing
information about the repositories belonging to that respective organization
}
\author{
Jan Philipp Dietrich
}
