% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incrementVersion.R
\name{incrementVersion}
\alias{incrementVersion}
\title{incrementVersion}
\usage{
incrementVersion(currentVersion, position, defLengths = 3)
}
\arguments{
\item{currentVersion}{The current package version as a string like "1.23.456"}

\item{position}{An integer defining which part of the version number will be increased. Use 1 for major version,
2 for minor, 3 for patch/bugfix, 4 for development.}

\item{defLengths}{An integer defining how many parts make up the resulting version number.}
}
\value{
The new version string.
}
\description{
Increment a version number at the specified position.
}
\examples{
lucode2:::incrementVersion("1.23.45", 3)
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich, Anastasis Giannousakis, Markus Bonsch, Pascal Sauer
}
