% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBuildLibraryConfig.R
\name{loadBuildLibraryConfig}
\alias{loadBuildLibraryConfig}
\title{buildLibrary}
\usage{
loadBuildLibraryConfig(lib)
}
\arguments{
\item{lib}{Path to the package}
}
\value{
The configuration loaded from .buildLibrary as a list.
}
\description{
Load the build configuration from .buildLibrary. If the file does not exist it is created.
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich, Pascal Führlich
}
