% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintain_libraries.R
\name{maintain_libraries}
\alias{maintain_libraries}
\title{Library cleanup tool}
\usage{
maintain_libraries(libpath = ".")
}
\arguments{
\item{libpath}{Parent path of the library folder}
}
\description{
Cleans up R library directory. Three steps are taken at the moment:
.Rcheck-folders are deleted, old packages are removed, missing zip or tar.gz
packages are created (for zip files a Windows system is required).
}
\examples{
\dontrun{
maintain_libraries()
}

}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich
}
