% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateConfig.R
\name{manipulateConfig}
\alias{manipulateConfig}
\title{Replace in File}
\usage{
manipulateConfig(configFile, ...)
}
\arguments{
\item{configFile}{a character string containing the name of the
configuration file, that should be manipulated. Supported file formats are
at the moment "gms", "inc", "cfg" (R-syntax), "php", "opt" and "cmd". Other
formats are currently not supported}

\item{...}{Variables, that should be set to new values, e.g. title="test"
for default.cfg or s_max_timesteps=10 for magpie.gms}
}
\description{
Function to set configuration parameters in configuration files (e.g.
default.cfg and magpie.gms). This replacement is useful, when using R to
manage different model runs at once.  Please check your results after
replacement!
}
\examples{
\dontrun{
manipulateConfig("config/default.cfg", input = "test_new_yields", title = "yihaa", revision = 4.2)
manipulateConfig("magpie.gms", s_max_timesteps = 4, s_use_gdx = -1)
}

}
\seealso{
\code{\link{manipulateFile}}
}
\author{
Jan Philipp Dietrich, Markus Bonsch, David Klein
}
