% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memCheck.R
\name{memCheck}
\alias{memCheck}
\title{Memory usage Check}
\usage{
memCheck(
  order.by = "Size",
  decreasing = TRUE,
  n = NULL,
  envir = parent.frame(),
  gc = TRUE
)
}
\arguments{
\item{order.by}{Column based on which the data should be sorted}

\item{decreasing}{Determines whether the values should be in an increasing
or decreasing order}

\item{n}{Limit of number of elements that should be shown. NULL means no
limit}

\item{envir}{Environment which should be analyzed, but default the parent
environment relative to this function.}

\item{gc}{Determines whether the garbage collector should be executed at the
end of the function for additional information}
}
\description{
Function checks memory usage and shows the biggest objects in the given
environment
}
\details{
This function is based on an idea posted at stack overflow:
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\examples{
\dontrun{
memCheck()
}
}
\author{
Jan Philipp Dietrich
}
