% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package2readme.R
\name{package2readme}
\alias{package2readme}
\title{package2readme}
\usage{
package2readme(package = ".")
}
\arguments{
\item{package}{either the path to the main folder of a package (containing a DESCRIPTION file)
or the name of the package}
}
\description{
Creates a README.md for a R package.
}
\examples{
package2readme("lucode2")
}
\author{
Jan Philipp Dietrich
}
