% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamPackages.R
\name{piamPackages}
\alias{piamPackages}
\title{piamPackages}
\usage{
piamPackages()
}
\value{
A character vector of names of packages available on https://pik-piam.r-universe.dev/ui#builds
}
\description{
Fetches the names of packages available on https://pik-piam.r-universe.dev/ui#builds
}
