% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_missing_reports.R
\name{produce_missing_reports}
\alias{produce_missing_reports}
\title{Produces the reporting mif files where they are missing}
\usage{
produce_missing_reports(modeldir = "./")
}
\arguments{
\item{modeldir}{Path to the main folder of REMIND or MAgPIE.}
}
\description{
For coupled runs: Searches the output folder for all existing run folders,
checks which of them are currently running on the cluster, ignores them,
checks for the remaining runs whether there is a reporting. Produces the
reporting if it is missing.
#'
}
\author{
David Klein
}
