% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArgs.R
\name{readArgs}
\alias{readArgs}
\title{Read Arguments from command line}
\usage{
readArgs(..., .envir = parent.frame(), .silent = FALSE)
}
\arguments{
\item{\dots}{arguments allowed to be read from command line (other values
are ignored). Value is set if found on command line input, nothing is done,
if value is not found.}

\item{.envir}{environment in which the variables should be written (by
default the environment from which the function is called)}

\item{.silent}{boolean which allows to suppress status messages}
}
\description{
Function reads arguments from command line of the structure value=content
and transforms them to R-Values, if they are called as allowed arguments.
}
\examples{

# Create an R-file "test.R" with following code:
value1 <- "old"
value2 <- 2
value3 <- "willstaythesame"
readArgs("value1", "value2", "value4")
cat(value1, "\n")
cat(value2, "\n")
cat(value3, "\n")

# Open the command line and execute the following code:
# Rscript test.R value1=new value2=3 value3=isnotallowed

# Output:
#
# ### READ COMMAND LINE - ASSIGNED CONFIGURATION ###
# value1 <- new
# value2 <- 3
# value4 not found
# ### READ COMMAND LINE - CONFIGURATION END ###
#
# "new"
# 3
# "willstaythesame"


### function that reads all allowed arguments from command line###
}
\seealso{
\code{\link{manipulateConfig}}
}
\author{
Jan Philipp Dietrich
}
