% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRuntime.R
\name{readRuntime}
\alias{readRuntime}
\title{readRuntime}
\usage{
readRuntime(path, plot = FALSE, types = NULL, coupled = FALSE, outfname = NULL)
}
\arguments{
\item{path}{Path to a run or a vector of paths.}

\item{plot}{Logical indicating whether the output should be plotted to a pdf
with the name runtime.pdf.}

\item{types}{A vector of names of different types of runs which should be
distinguished by colors. The names have to be part of the folder name in
order to allow the function to map the given types to the runs.}

\item{coupled}{Logical indicating if comparison plots should be added for
coupled REMIND and MAgPIE runs. \code{TRUE} automatically sets \code{types}
to \code{c("-rem-","-mag-")} and overwrites user defined \code{types}}

\item{outfname}{Optional name of the pdf. If nothing is given the default
"runtime" will be used.}
}
\value{
A data frame containing the run names and runtime information in
hours.
}
\description{
Reads all runtime information from given experiments. The runtime is given
in hours and is the runtime of GAMS.
}
\author{
David Klein
}
