% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEOF.R
\name{removeEOF}
\alias{removeEOF}
\title{add EOF comment}
\usage{
removeEOF(path = ".", filetypes = c("inc", "prn", "gms"))
}
\arguments{
\item{path}{path to the main folder of the model}

\item{filetypes}{file types the function should be applied to}
}
\description{
remove EOF text from all files in the code
}
\examples{
\dontrun{
removeEOF()
}

}
\seealso{
\code{\link{addEOF}}
}
\author{
Anastasis Giannousakis
}
