% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmAllbut.R
\name{rmAllbut}
\alias{rmAllbut}
\title{rmAllbut}
\usage{
rmAllbut(..., list = character(), clean = TRUE)
}
\arguments{
\item{...}{Objects that should be kept}

\item{list}{List specifying the objects to be kept. Same as in \code{\link[base]{rm}}.}

\item{clean}{Boolean to specify if a \code{\link[base]{gc}} shall be executed}
}
\description{
Removes all objects except specified ones from the workspace
}
\details{
Helps to clean the workspace. Only objects specified in \code{...} survive.
Specify clean =TRUE to really free the memory.
}
\examples{

# Create some objects
a <- 1
b <- 2
c <- 3
# show them
ls()
# delete all but b and c
rmAllbut(b, c)
ls()
# delete all but b
test <- "b"
rmAllbut(list = test)
ls()
}
\seealso{
\code{\link[base]{rm}},\code{\link[base]{ls}}
}
\author{
Markus Bonsch
}
