% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstatistics.R
\name{runstatistics}
\alias{runstatistics}
\title{Run Statistics}
\usage{
runstatistics(file = "runstatistics.Rda", overwrite = TRUE, submit = NULL, ...)
}
\arguments{
\item{file}{file name the statistics are/should be stored}

\item{overwrite}{boolean deciding whether entries should be overwritten, if already existing
in the file. If set to FALSE an error will be thrown in case that an overwrite is attempted.}

\item{submit}{path to a folder the run statistics should be submitted to. As soon as the path is
set the data will be submitted, so please only set the path as soon as the run statistics are
complete}

\item{...}{entries that should be added to the run statistics file. Standard entries are:
model, config, runtime, user, date, modelstat, version_management, revision and status.}
}
\value{
An invisible list containing run statistics as stored in the given file
}
\description{
Support function to collect run statistics.
}
\examples{
f <- tempfile()
 runstatistics(file = f, user = Sys.info()[["user"]])
 print(runstatistics(file = f))
 runstatistics(file = f, submit = tempdir())
}
\author{
Jan Philipp Dietrich
}
