% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendmail.R
\name{sendmail}
\alias{sendmail}
\title{sendmail}
\usage{
sendmail(
  path = NULL,
  gitrepo,
  file,
  commitmessage,
  remote = FALSE,
  reset = FALSE
)
}
\arguments{
\item{path}{path to the clone of the gitlab repo (can be NULL)}

\item{gitrepo}{if no path is given, the gitlab repo is needed}

\item{file}{absolute path to the file to be committed}

\item{commitmessage}{the commit message (appears in the subject line of the email that will be sent)}

\item{remote}{whether communication with a remote is needed}

\item{reset}{whether a reset of a local copy is wanted}
}
\description{
A function that sends an automatic email with each push to a gitlab repo
}
\author{
Anastasis Giannousakis
}
