% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_info.R
\name{setup_info}
\alias{setup_info}
\title{Setup Info}
\usage{
setup_info()
}
\value{
A list with information about the current session and the currently
used R setup.
}
\description{
Returns a list with information about the current session (session info,
used library paths and installed libraries)
}
\examples{

setup_info()
}
\author{
Jan Philipp Dietrich
}
