% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd2.R
\name{setwd2}
\alias{setwd2}
\title{setwd2}
\usage{
setwd2(return_only = FALSE)
}
\arguments{
\item{return_only}{if TRUE, the path is not changed, but the clipboard path is returned as string.}
}
\value{
if return_only=FALSE: Nothing, but the working directory is set to. Otherwise: no working directory returned,
        but path transformed.
}
\description{
Mini function that allows you to set a directory based on a readline input.
Very useful for Windows users, as it replaces backslashes by slashes.
}
\author{
Benjamin Leon Bodirsky
}
