% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakeToCamel.R
\name{snakeToCamel}
\alias{snakeToCamel}
\title{snakeToCamel}
\usage{
snakeToCamel(pathToFile, ask = TRUE)
}
\arguments{
\item{pathToFile}{Path to the R source file where variables should be renamed.}

\item{ask}{If TRUE (default) ask before renaming a variable, otherwise always assume "yes" as answer.}
}
\description{
Convenience function to rename variables in an R file from snake to camel case.
}
