% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemCommandAvailable.R
\name{SystemCommandAvailable}
\alias{SystemCommandAvailable}
\title{System Command Available}
\usage{
SystemCommandAvailable(command)
}
\arguments{
\item{command}{System command to be tested}
}
\value{
Boolean indicating whether the command is available, or not
}
\description{
Checks whether a system command is available (does not return an error), or not
}
\examples{
SystemCommandAvailable("ls")
}
\author{
Jan Philipp Dietrich
}
