% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPackage.R
\name{testPackage}
\alias{testPackage}
\title{Test package}
\usage{
testPackage(repo, tmpLib = tempdir(), ...)
}
\arguments{
\item{repo}{GitHub repository to install the package from}

\item{tmpLib}{temporary library directory where the package should be installed}

\item{...}{additional arguments forwarded to \code{devtools::install_github}}
}
\description{
Installs a package in a temporary library and loads that library on top of the existing one
}
\examples{
\dontrun{
testPackage("git@github.com:pik-piam/lucode2")
}

}
\author{
Jan Philipp Dietrich
}
