% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRepo.R
\name{updateRepo}
\alias{updateRepo}
\title{Update package repository}
\usage{
updateRepo(
  path = ".",
  check = TRUE,
  forceRebuild = FALSE,
  clean = FALSE,
  pidfile = NULL
)
}
\arguments{
\item{path}{Path to the repository}

\item{check}{Boolean deciding whether package must have been checked or not
in order to be distributed}

\item{forceRebuild}{Option to rebuild all packages from source}

\item{clean}{Option to clean repos before updating/pulling to avoid merge conflicts}

\item{pidfile}{file name of a "process id" file containing the process id of the current
process. If set this will make sure that the function is not run twice at the same time.
If set to NULL this will be ignored. The OS needs to provide a "ps" command in order to
have this working.}
}
\description{
Function to update an package repository. Run this function on a folder which contains
packages sources as subfolders. Packages should be managed via git in order to be updated properly.
To add a new package to the repo just checkout/clone it into this folder. The
function will automatically detect the new package and add it.
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich
}
