% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationkey.R
\name{validationkey}
\alias{validationkey}
\title{validationkey}
\usage{
validationkey(version, date)
}
\arguments{
\item{version}{Version number of the package}

\item{date}{Date of the package}
}
\description{
Support function which creates a key out of a version date combination
}
\details{
This function is used in \code{\link{buildLibrary}} to offer the package
publication server an option to check whether the package has been properly
and successfully (no errors/warnings/notes) checked before its commit.

The calculated key is not safe and can easily be reproduced, but should be
complicated enough to encourage users running the \code{\link{buildLibrary}}
check properly.
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich
}
