% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validkey.R
\name{validkey}
\alias{validkey}
\title{validkey}
\usage{
validkey(package = ".")
}
\arguments{
\item{package}{Path to the package}
}
\value{
list with version, date and result of validation test
}
\description{
Support function which validates a key out of a version date combination
}
\details{
This function is used to check whether \code{\link{buildLibrary}} has been run
properly and without problems or not
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich
}
