% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{verifyCheck}
\alias{verifyCheck}
\title{verifyCheck}
\usage{
verifyCheck(cran, acceptedWarnings, acceptedNotes)
}
\arguments{
\item{cran}{Passed to devtools::check}

\item{acceptedWarnings}{A character vector of regular expressions.
A warning will result in an error unless it matches one of these regular expressions.}

\item{acceptedNotes}{A character vector of regular expressions.
A note will result in an error unless it matches one of these regular expressions.}
}
\description{
Run R CMD check completely without stopping. Then stop on errors, or unaccepted warnings and notes.
}
\author{
Pascal Sauer
}
