% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{verifyLinter}
\alias{verifyLinter}
\title{verifyLinter}
\usage{
verifyLinter(allowLinterWarnings = FALSE)
}
\arguments{
\item{allowLinterWarnings}{If FALSE (the default) will stop on linter warnings.}
}
\description{
Run linter and stop on linter warning unless linter warnings are allowed.
}
\author{
Pascal Sauer
}
