% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{verifyTests}
\alias{verifyTests}
\title{verifyTests}
\usage{
verifyTests(acceptedWarnings)
}
\arguments{
\item{acceptedWarnings}{A character vector of regular expressions.
A warning will result in an error unless it matches one of these regular expressions.}
}
\description{
Run tests and stop on error or unaccepted warning.
}
\author{
Pascal Sauer
}
