% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGitHubActions.R
\name{addGitHubActions}
\alias{addGitHubActions}
\title{addGitHubActions}
\usage{
addGitHubActions(lib = ".")
}
\arguments{
\item{lib}{Path to the package}
}
\description{
adds GitHubActions with a standard workflow including codecov to the package. It will add a standard
Github action "test-buildlibrary.yml" to the project which performs standard tests as well as
creates a coverage report. This file is overwritten automatically each time this function is run and
should not be edited by hand. But additional Github actions can be added as separate files.
}
\details{
In addition, this function adds a codecov.yml to the repository, if not already existing. This file
is only created, if missing and can be edited manually.
}
\examples{
\dontrun{
addGitHubActions()
}
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich
}
