% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFormat.R
\name{autoFormat}
\alias{autoFormat}
\title{autoFormat}
\usage{
autoFormat(
  files = getFilesToLint(),
  ignoreLintFreeFiles = TRUE,
  lintAfterwards = TRUE
)
}
\arguments{
\item{files}{A character vector of paths to files that should be auto-formatted.}

\item{ignoreLintFreeFiles}{If set to TRUE (the default) files without linter warnings are not auto-formatted.}

\item{lintAfterwards}{If set to TRUE (the default) return linter results for the auto-formatted files.}
}
\description{
Apply auto-formatting using styler::style_file to the given files. Does not change indentation.
Use autoFormat to solve many linter warning at once, but only those which are safe to solve
automatically. Also check the function formatTF and snakeToCamel for assistance in solving
specific warnings.
}
\examples{
\dontrun{
lucode2::autoFormat()
}
}
\seealso{
\code{\link{getFilesToLint}}, \code{\link{formatTF}}, \code{\link{snakeToCamel}}
}
\author{
Pascal Führlich
}
