% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLibrary.R
\name{buildLibrary}
\alias{buildLibrary}
\title{buildLibrary}
\usage{
buildLibrary(
  lib = ".",
  cran = TRUE,
  updateType = NULL,
  gitpush = FALSE,
  commitmessage = NULL
)
}
\arguments{
\item{lib}{Path to the package}

\item{cran}{If cran-like test is needed}

\item{updateType}{Either an integer or character string:\tabular{lll}{
   \strong{number} \tab \strong{string} \tab \strong{description} \cr
   1 \tab \code{major} \tab for major rewrite of the whole package \cr
   2 (default) \tab \code{minor} \tab for new features or improvements \cr
   3 \tab \code{patch} \tab for bugfixes and corrections \cr
   4 \tab \code{development} \tab only for packages in development stage \cr
}}

\item{gitpush}{If a git commit should happen automatically}

\item{commitmessage}{Your commit message}
}
\description{
Builds R libraries. Includes checks for consistency.
}
\details{
This function is designed to help building R libraries. It performs the
following steps: \itemize{
\item Version: Determination of a new version number (Can also be defined by the user).
\item Date: Determination of a the date of the build (Can also be defined by the user).
\item Linter: Check for code style problems.
\item R check: Check whether the library is consistent and can be built.
\item Package building: Builds the .zip and .tar.gz packages under windows.
Under linux, only the .tar.gz package is built. }
The commit has to be performed by the user still.
}
\examples{
\dontrun{
buildLibrary()
}
}
\seealso{
\code{\link{package2readme}}, \code{\link{lint}}, \code{\link{autoFormat}}
}
\author{
Jan Philipp Dietrich, Anastasis Giannousakis, Markus Bonsch, Pascal Führlich
}
