% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{check}
\usage{
check(
  lib = ".",
  cran = TRUE,
  config = loadBuildLibraryConfig(lib),
  runLinter = TRUE
)
}
\arguments{
\item{lib}{Path to the package}

\item{cran}{If cran-like test is needed}

\item{config}{A configuration defining AcceptedWarnings, AcceptedNotes, and allowLinterWarnings. By default the
.buildLibrary file is read.}

\item{runLinter}{Set to FALSE to skip the linter.}
}
\description{
Runs devtools::test(), devtools::check() (without tests), and lucode2::lint()
}
\examples{
\dontrun{
lucode2::check()
}
}
\seealso{
\code{\link{buildLibrary}}, \code{\link{lint}}

\code{\link[devtools]{check}}, \code{\link[devtools]{test}}
}
\author{
Jan Philipp Dietrich, Pascal Führlich
}
