% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVersionUpdated.R
\name{isVersionUpdated}
\alias{isVersionUpdated}
\title{isVersionUpdated}
\usage{
isVersionUpdated(repo = "https://rse.pik-potsdam.de/r/packages/")
}
\arguments{
\item{repo}{package repository to determine latest version}
}
\description{
Checks if the version number in the DESCRIPTION file of a given package has
been updated.
}
\author{
Falk Benke
}
