% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVersionUpdated.R
\name{isVersionUpdated}
\alias{isVersionUpdated}
\title{isVersionUpdated}
\usage{
isVersionUpdated(
  repo = "https://rse.pik-potsdam.de/r/packages/",
  config = loadBuildLibraryConfig()
)
}
\arguments{
\item{repo}{package repository to determine latest version}

\item{config}{A configuration defining enforceVersionUpdate. By default
the .buildLibrary file is read.}
}
\description{
Checks if the version number in the DESCRIPTION file of a given package has
been updated (may not be a version number for development stage packages).
}
\author{
Falk Benke
}
