% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint}
\alias{lint}
\title{lint}
\usage{
lint(files = getFilesToLint())
}
\arguments{
\item{files}{A character vector of paths to files that should be checked by the linter. If set to "."
the whole package is linted.}
}
\value{
A named list, where the names are the paths to the linted files and the values are lists containing the
linter warnings.
}
\description{
Check the given files for linter warnings using lintr::lint.
}
\details{
For files in the vignettes and tests folder less strict rules are applied, e.g. using ::: usually
leads to a linter warning, but not in vignettes/tests. Which linter rules are used depends on
".lintr" config files. \code{\link{check}} creates lintr config files that use \code{\link{lintrRules}}.
}
\examples{
\dontrun{
lucode2::lint()
}
}
\seealso{
\code{\link{getFilesToLint}}, \code{\link{lintrRules}}, \code{\link{autoFormat}}, \code{\link[lintr]{lint}}
}
\author{
Pascal Führlich
}
