% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintrRules.R
\name{lintrRules}
\alias{lintrRules}
\title{lintrRules}
\usage{
lintrRules(allowUndesirable = FALSE)
}
\arguments{
\item{allowUndesirable}{If true it is okay to use undesirable operators (such as "<<-")
and undesirable (but not deprecated) functions (such as "setwd").}
}
\description{
This function defines the rules to be used by the linter called lintr. \code{\link{check}} creates ".lintr"
config files that use this function.
}
\examples{
\dontrun{lintr::lint_dir(linters = lintrRules())}

}
\seealso{
\code{\link{check}}, \code{\link{lint}}
}
