% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucode2-package.R
\docType{package}
\name{lucode2-package}
\alias{lucode2}
\alias{lucode2-package}
\title{lucode2: Code Manipulation and Analysis Tools}
\description{
A collection of tools which allow to manipulate and analyze code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/lucode2}
  \item \doi{10.5281/zenodo.4389418}
  \item Report bugs at \url{https://github.com/pik-piam/lucode2/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Pascal Führlich
  \item David Klein
  \item Anastasis Giannousakis
  \item Markus Bonsch
  \item Benjamin Leon Bodirsky
  \item Lavinia Baumstark
  \item Oliver Richters
}

}
