% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package2readme.R
\name{package2readme}
\alias{package2readme}
\title{package2readme}
\usage{
package2readme(package = ".", add = NULL, logo = NULL)
}
\arguments{
\item{package}{either the path to the main folder of a package (containing a DESCRIPTION file)
or the name of the package}

\item{add}{a character vector with additions to the README file. Each element of the vector can be either
1) a line of markdown code, 2) a path to a markdown file, or 3) a path to a Rmarkdown file}

\item{logo}{a character string for a path to a logo file used in the title of the README file}
}
\description{
Creates a README.md for a R package.
}
\examples{
package2readme("lucode2")
}
\author{
Jan Philipp Dietrich
}
