% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRepo.R
\name{updateRepo}
\alias{updateRepo}
\title{Update package repository}
\usage{
updateRepo(
  path = ".",
  check = TRUE,
  forceRebuild = FALSE,
  clean = FALSE,
  skipFolders = c("Archive", "gdxrrw", "HARr"),
  repoUrl = "https://rse.pik-potsdam.de/r/packages"
)
}
\arguments{
\item{path}{Path to the repository}

\item{check}{Boolean deciding whether package must have been checked or not in order to be distributed}

\item{forceRebuild}{Option to rebuild all packages from source}

\item{clean}{Option to clean repos before updating/pulling to avoid merge conflicts}

\item{skipFolders}{Which folders/packages should not be built.}

\item{repoUrl}{Url of the package repo. Will be added to DESCRIPTION files in the Repository field.}
}
\description{
Function to update an package repository. Run this function on a folder which contains
packages sources as subfolders. Packages should be managed via git in order to be updated properly.
To add a new package to the repo just checkout/clone it into this folder. The
function will automatically detect the new package and add it.
}
\seealso{
\code{\link{buildLibrary}}
}
\author{
Jan Philipp Dietrich, Pascal Führlich
}
